requirements.txt
hypothesis
matplotlib
numpy
opencv_python
Pillow
setuptools
